Forces navController.navigateWithId to work with Long instead of Int

+++ Start/app/src/main/java/com/esei/grvidal/nighttime/pages/Friends.kt
@@ -36,7 +36,7 @@ fun FriendsPageView(navController: NavHostController) {
         LazyColumnFor(items = user.getChats()) {
 
             ChatEntry(userName = it.userName, lastMessage = it.lastMessage,
-                onEntryClick = { navController.navigateWithId( NavigationScreens.ChatConversation.route, it.id ) }
+                onEntryClick = { navController.navigateWithId( NavigationScreens.ChatConversation.route, it.id.toLong() ) }
             )
         }
     }
